Module aws_sdk_appmesh::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

internal_server_error_exception

See InternalServerErrorException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

resource_in_use_exception

See ResourceInUseException

service_unavailable_exception

See ServiceUnavailableException

too_many_requests_exception

See TooManyRequestsException

too_many_tags_exception

See TooManyTagsException

Structs

BadRequestException

The request syntax was malformed. Check your request syntax and try again.

ConflictException

The request contains a client token that was used for a previous update resource call

CreateGatewayRouteError

CreateMeshError

CreateRouteError

CreateVirtualGatewayError

CreateVirtualNodeError

CreateVirtualRouterError

CreateVirtualServiceError

DeleteGatewayRouteError

DeleteMeshError

DeleteRouteError

DeleteVirtualGatewayError

DeleteVirtualNodeError

DeleteVirtualRouterError

DeleteVirtualServiceError

DescribeGatewayRouteError

DescribeMeshError

DescribeRouteError

DescribeVirtualGatewayError

DescribeVirtualNodeError

DescribeVirtualRouterError

DescribeVirtualServiceError

ForbiddenException

You don't have permissions to perform this action.

InternalServerErrorException

The request processing has failed because of an unknown error, exception, or

LimitExceededException

You have exceeded a service limit for your account. For more information, see Service

ListGatewayRoutesError

ListMeshesError

ListRoutesError

ListTagsForResourceError

ListVirtualGatewaysError

ListVirtualNodesError

ListVirtualRoutersError

ListVirtualServicesError

NotFoundException

The specified resource doesn't exist. Check your request syntax and try again.

ResourceInUseException

You can't delete the specified resource because it's in use or required by another

ServiceUnavailableException

The request has failed due to a temporary failure of the service.

TagResourceError

TooManyRequestsException

The maximum request rate permitted by the App Mesh APIs has been exceeded for your

TooManyTagsException

The request exceeds the maximum allowed number of tags allowed per resource. The current

UntagResourceError

UpdateGatewayRouteError

UpdateMeshError

UpdateRouteError

UpdateVirtualGatewayError

UpdateVirtualNodeError

UpdateVirtualRouterError

UpdateVirtualServiceError

Enums

CreateGatewayRouteErrorKind

CreateMeshErrorKind

CreateRouteErrorKind

CreateVirtualGatewayErrorKind

CreateVirtualNodeErrorKind

CreateVirtualRouterErrorKind

CreateVirtualServiceErrorKind

DeleteGatewayRouteErrorKind

DeleteMeshErrorKind

DeleteRouteErrorKind

DeleteVirtualGatewayErrorKind

DeleteVirtualNodeErrorKind

DeleteVirtualRouterErrorKind

DeleteVirtualServiceErrorKind

DescribeGatewayRouteErrorKind

DescribeMeshErrorKind

DescribeRouteErrorKind

DescribeVirtualGatewayErrorKind

DescribeVirtualNodeErrorKind

DescribeVirtualRouterErrorKind

DescribeVirtualServiceErrorKind

ListGatewayRoutesErrorKind

ListMeshesErrorKind

ListRoutesErrorKind

ListTagsForResourceErrorKind

ListVirtualGatewaysErrorKind

ListVirtualNodesErrorKind

ListVirtualRoutersErrorKind

ListVirtualServicesErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateGatewayRouteErrorKind

UpdateMeshErrorKind

UpdateRouteErrorKind

UpdateVirtualGatewayErrorKind

UpdateVirtualNodeErrorKind

UpdateVirtualRouterErrorKind

UpdateVirtualServiceErrorKind