Module aws_sdk_appmesh::model

Modules

aws_cloud_map_instance_attribute

See AwsCloudMapInstanceAttribute

aws_cloud_map_service_discovery

See AwsCloudMapServiceDiscovery

backend_defaults

See BackendDefaults

client_policy

See ClientPolicy

client_policy_tls

See ClientPolicyTls

dns_service_discovery

See DnsServiceDiscovery

duration

See Duration

egress_filter

See EgressFilter

file_access_log

See FileAccessLog

gateway_route_data

See GatewayRouteData

gateway_route_hostname_match

See GatewayRouteHostnameMatch

gateway_route_hostname_rewrite

See GatewayRouteHostnameRewrite

gateway_route_ref

See GatewayRouteRef

gateway_route_spec

See GatewayRouteSpec

gateway_route_status

See GatewayRouteStatus

gateway_route_target

See GatewayRouteTarget

gateway_route_virtual_service

See GatewayRouteVirtualService

grpc_gateway_route

See GrpcGatewayRoute

grpc_gateway_route_action

See GrpcGatewayRouteAction

grpc_gateway_route_match

See GrpcGatewayRouteMatch

grpc_gateway_route_metadata

See GrpcGatewayRouteMetadata

grpc_gateway_route_rewrite

See GrpcGatewayRouteRewrite

grpc_retry_policy

See GrpcRetryPolicy

grpc_route

See GrpcRoute

grpc_route_action

See GrpcRouteAction

grpc_route_match

See GrpcRouteMatch

grpc_route_metadata

See GrpcRouteMetadata

grpc_timeout

See GrpcTimeout

health_check_policy

See HealthCheckPolicy

http_gateway_route

See HttpGatewayRoute

http_gateway_route_action

See HttpGatewayRouteAction

http_gateway_route_header

See HttpGatewayRouteHeader

http_gateway_route_match

See HttpGatewayRouteMatch

http_gateway_route_path_rewrite

See HttpGatewayRoutePathRewrite

http_gateway_route_prefix_rewrite

See HttpGatewayRoutePrefixRewrite

http_gateway_route_rewrite

See HttpGatewayRouteRewrite

http_path_match

See HttpPathMatch

http_query_parameter

See HttpQueryParameter

http_retry_policy

See HttpRetryPolicy

http_route

See HttpRoute

http_route_action

See HttpRouteAction

http_route_header

See HttpRouteHeader

http_route_match

See HttpRouteMatch

http_timeout

See HttpTimeout

listener

See Listener

listener_tls

See ListenerTls

listener_tls_acm_certificate

See ListenerTlsAcmCertificate

listener_tls_file_certificate

See ListenerTlsFileCertificate

listener_tls_sds_certificate

See ListenerTlsSdsCertificate

listener_tls_validation_context

See ListenerTlsValidationContext

logging

See Logging

match_range

See MatchRange

mesh_data

See MeshData

mesh_ref

See MeshRef

mesh_spec

See MeshSpec

mesh_status

See MeshStatus

outlier_detection

See OutlierDetection

port_mapping

See PortMapping

query_parameter_match

See QueryParameterMatch

resource_metadata

See ResourceMetadata

route_data

See RouteData

route_ref

See RouteRef

route_spec

See RouteSpec

route_status

See RouteStatus

subject_alternative_name_matchers

See SubjectAlternativeNameMatchers

subject_alternative_names

See SubjectAlternativeNames

tag_ref

See TagRef

tcp_route

See TcpRoute

tcp_route_action

See TcpRouteAction

tcp_timeout

See TcpTimeout

tls_validation_context

See TlsValidationContext

tls_validation_context_acm_trust

See TlsValidationContextAcmTrust

tls_validation_context_file_trust

See TlsValidationContextFileTrust

tls_validation_context_sds_trust

See TlsValidationContextSdsTrust

virtual_gateway_backend_defaults

See VirtualGatewayBackendDefaults

virtual_gateway_client_policy

See VirtualGatewayClientPolicy

virtual_gateway_client_policy_tls

See VirtualGatewayClientPolicyTls

virtual_gateway_data

See VirtualGatewayData

virtual_gateway_file_access_log

See VirtualGatewayFileAccessLog

virtual_gateway_grpc_connection_pool

See VirtualGatewayGrpcConnectionPool

virtual_gateway_health_check_policy

See VirtualGatewayHealthCheckPolicy

virtual_gateway_http2_connection_pool

See VirtualGatewayHttp2ConnectionPool

virtual_gateway_http_connection_pool

See VirtualGatewayHttpConnectionPool

virtual_gateway_listener

See VirtualGatewayListener

virtual_gateway_listener_tls

See VirtualGatewayListenerTls

virtual_gateway_listener_tls_acm_certificate

See VirtualGatewayListenerTlsAcmCertificate

virtual_gateway_listener_tls_file_certificate

See VirtualGatewayListenerTlsFileCertificate

virtual_gateway_listener_tls_sds_certificate

See VirtualGatewayListenerTlsSdsCertificate

virtual_gateway_listener_tls_validation_context

See VirtualGatewayListenerTlsValidationContext

virtual_gateway_logging

See VirtualGatewayLogging

virtual_gateway_port_mapping

See VirtualGatewayPortMapping

virtual_gateway_ref

See VirtualGatewayRef

virtual_gateway_spec

See VirtualGatewaySpec

virtual_gateway_status

See VirtualGatewayStatus

virtual_gateway_tls_validation_context

See VirtualGatewayTlsValidationContext

virtual_gateway_tls_validation_context_acm_trust

See VirtualGatewayTlsValidationContextAcmTrust

virtual_gateway_tls_validation_context_file_trust

See VirtualGatewayTlsValidationContextFileTrust

virtual_gateway_tls_validation_context_sds_trust

See VirtualGatewayTlsValidationContextSdsTrust

virtual_node_data

See VirtualNodeData

virtual_node_grpc_connection_pool

See VirtualNodeGrpcConnectionPool

virtual_node_http2_connection_pool

See VirtualNodeHttp2ConnectionPool

virtual_node_http_connection_pool

See VirtualNodeHttpConnectionPool

virtual_node_ref

See VirtualNodeRef

virtual_node_service_provider

See VirtualNodeServiceProvider

virtual_node_spec

See VirtualNodeSpec

virtual_node_status

See VirtualNodeStatus

virtual_node_tcp_connection_pool

See VirtualNodeTcpConnectionPool

virtual_router_data

See VirtualRouterData

virtual_router_listener

See VirtualRouterListener

virtual_router_ref

See VirtualRouterRef

virtual_router_service_provider

See VirtualRouterServiceProvider

virtual_router_spec

See VirtualRouterSpec

virtual_router_status

See VirtualRouterStatus

virtual_service_backend

See VirtualServiceBackend

virtual_service_data

See VirtualServiceData

virtual_service_ref

See VirtualServiceRef

virtual_service_spec

See VirtualServiceSpec

virtual_service_status

See VirtualServiceStatus

weighted_target

See WeightedTarget

Structs

AwsCloudMapInstanceAttribute

An object that represents the Cloud Map attribute information for your virtual

AwsCloudMapServiceDiscovery

An object that represents the Cloud Map service discovery information for your virtual

BackendDefaults

An object that represents the default properties for a backend.

ClientPolicy

An object that represents a client policy.

ClientPolicyTls

A reference to an object that represents a Transport Layer Security (TLS) client policy.

DnsServiceDiscovery

An object that represents the DNS service discovery information for your virtual

Duration

An object that represents a duration of time.

EgressFilter

An object that represents the egress filter rules for a service mesh.

FileAccessLog

An object that represents an access log file.

GatewayRouteData

An object that represents a gateway route returned by a describe operation.

GatewayRouteHostnameMatch

An object representing the gateway route host name to match.

GatewayRouteHostnameRewrite

An object representing the gateway route host name to rewrite.

GatewayRouteRef

An object that represents a gateway route returned by a list operation.

GatewayRouteSpec

An object that represents a gateway route specification. Specify one gateway route

GatewayRouteStatus

An object that represents the current status of a gateway route.

GatewayRouteTarget

An object that represents a gateway route target.

GatewayRouteVirtualService

An object that represents the virtual service that traffic is routed to.

GrpcGatewayRoute

An object that represents a gRPC gateway route.

GrpcGatewayRouteAction

An object that represents the action to take if a match is determined.

GrpcGatewayRouteMatch

An object that represents the criteria for determining a request match.

GrpcGatewayRouteMetadata

An object representing the metadata of the gateway route.

GrpcGatewayRouteRewrite

An object that represents the gateway route to rewrite.

GrpcRetryPolicy

An object that represents a retry policy. Specify at least one value for at least one of the types of RetryEvents, a value for maxRetries, and a value for perRetryTimeout.

GrpcRoute

An object that represents a gRPC route type.

GrpcRouteAction

An object that represents the action to take if a match is determined.

GrpcRouteMatch

An object that represents the criteria for determining a request match.

GrpcRouteMetadata

An object that represents the match metadata for the route.

GrpcTimeout

An object that represents types of timeouts.

HealthCheckPolicy

An object that represents the health check policy for a virtual node's listener.

HttpGatewayRoute

An object that represents an HTTP gateway route.

HttpGatewayRouteAction

An object that represents the action to take if a match is determined.

HttpGatewayRouteHeader

An object that represents the HTTP header in the gateway route.

HttpGatewayRouteMatch

An object that represents the criteria for determining a request match.

HttpGatewayRoutePathRewrite

An object that represents the path to rewrite.

HttpGatewayRoutePrefixRewrite

An object representing the beginning characters of the route to rewrite.

HttpGatewayRouteRewrite

An object representing the gateway route to rewrite.

HttpPathMatch

An object representing the path to match in the request.

HttpQueryParameter

An object that represents the query parameter in the request.

HttpRetryPolicy

An object that represents a retry policy. Specify at least one value for at least one of the types of RetryEvents, a value for maxRetries, and a value for perRetryTimeout.

HttpRoute

An object that represents an HTTP or HTTP/2 route type.

HttpRouteAction

An object that represents the action to take if a match is determined.

HttpRouteHeader

An object that represents the HTTP header in the request.

HttpRouteMatch

An object that represents the requirements for a route to match HTTP requests for a

HttpTimeout

An object that represents types of timeouts.

Listener

An object that represents a listener for a virtual node.

ListenerTls

An object that represents the Transport Layer Security (TLS) properties for a listener.

ListenerTlsAcmCertificate

An object that represents an AWS Certicate Manager (ACM) certificate.

ListenerTlsFileCertificate

An object that represents a local file certificate.

ListenerTlsSdsCertificate

An object that represents the listener's Secret Discovery Service certificate. The proxy

ListenerTlsValidationContext

An object that represents a listener's Transport Layer Security (TLS) validation context.

Logging

An object that represents the logging information for a virtual node.

MatchRange

An object that represents the range of values to match on. The first character of the range is included in the range, though the last character is not. For example, if the range specified were 1-100, only values 1-99 would be matched.

MeshData

An object that represents a service mesh returned by a describe operation.

MeshRef

An object that represents a service mesh returned by a list operation.

MeshSpec

An object that represents the specification of a service mesh.

MeshStatus

An object that represents the status of a service mesh.

OutlierDetection

An object that represents the outlier detection for a virtual node's listener.

PortMapping

An object that represents a port mapping.

QueryParameterMatch

An object representing the query parameter to match.

ResourceMetadata

An object that represents metadata for a resource.

RouteData

An object that represents a route returned by a describe operation.

RouteRef

An object that represents a route returned by a list operation.

RouteSpec

An object that represents a route specification. Specify one route type.

RouteStatus

An object that represents the current status of a route.

SubjectAlternativeNameMatchers

An object that represents the methods by which a subject alternative name on a peer

SubjectAlternativeNames

An object that represents the subject alternative names secured by the

TagRef

Optional metadata that you apply to a resource to assist with categorization and

TcpRoute

An object that represents a TCP route type.

TcpRouteAction

An object that represents the action to take if a match is determined.

TcpTimeout

An object that represents types of timeouts.

TlsValidationContext

An object that represents how the proxy will validate its peer during Transport Layer Security (TLS)

TlsValidationContextAcmTrust

An object that represents a Transport Layer Security (TLS) validation context trust for an Certificate Manager

TlsValidationContextFileTrust

An object that represents a Transport Layer Security (TLS) validation context trust for a local file.

TlsValidationContextSdsTrust

An object that represents a Transport Layer Security (TLS) Secret Discovery Service validation context trust. The

VirtualGatewayBackendDefaults

An object that represents the default properties for a backend.

VirtualGatewayClientPolicy

An object that represents a client policy.

VirtualGatewayClientPolicyTls

An object that represents a Transport Layer Security (TLS) client policy.

VirtualGatewayData

An object that represents a virtual gateway returned by a describe operation.

VirtualGatewayFileAccessLog

An object that represents an access log file.

VirtualGatewayGrpcConnectionPool

An object that represents a type of connection pool.

VirtualGatewayHealthCheckPolicy

An object that represents the health check policy for a virtual gateway's

VirtualGatewayHttp2ConnectionPool

An object that represents a type of connection pool.

VirtualGatewayHttpConnectionPool

An object that represents a type of connection pool.

VirtualGatewayListener

An object that represents a listener for a virtual gateway.

VirtualGatewayListenerTls

An object that represents the Transport Layer Security (TLS) properties for a listener.

VirtualGatewayListenerTlsAcmCertificate

An object that represents an Certificate Manager certificate.

VirtualGatewayListenerTlsFileCertificate

An object that represents a local file certificate.

VirtualGatewayListenerTlsSdsCertificate

An object that represents the virtual gateway's listener's Secret Discovery Service

VirtualGatewayListenerTlsValidationContext

An object that represents a virtual gateway's listener's Transport Layer Security (TLS) validation

VirtualGatewayLogging

An object that represents logging information.

VirtualGatewayPortMapping

An object that represents a port mapping.

VirtualGatewayRef

An object that represents a virtual gateway returned by a list operation.

VirtualGatewaySpec

An object that represents the specification of a service mesh resource.

VirtualGatewayStatus

An object that represents the status of the mesh resource.

VirtualGatewayTlsValidationContext

An object that represents a Transport Layer Security (TLS) validation context.

VirtualGatewayTlsValidationContextAcmTrust

An object that represents a Transport Layer Security (TLS) validation context trust for an Certificate Manager

VirtualGatewayTlsValidationContextFileTrust

An object that represents a Transport Layer Security (TLS) validation context trust for a local file.

VirtualGatewayTlsValidationContextSdsTrust

An object that represents a virtual gateway's listener's Transport Layer Security (TLS) Secret Discovery Service

VirtualNodeData

An object that represents a virtual node returned by a describe operation.

VirtualNodeGrpcConnectionPool

An object that represents a type of connection pool.

VirtualNodeHttp2ConnectionPool

An object that represents a type of connection pool.

VirtualNodeHttpConnectionPool

An object that represents a type of connection pool.

VirtualNodeRef

An object that represents a virtual node returned by a list operation.

VirtualNodeServiceProvider

An object that represents a virtual node service provider.

VirtualNodeSpec

An object that represents the specification of a virtual node.

VirtualNodeStatus

An object that represents the current status of the virtual node.

VirtualNodeTcpConnectionPool

An object that represents a type of connection pool.

VirtualRouterData

An object that represents a virtual router returned by a describe operation.

VirtualRouterListener

An object that represents a virtual router listener.

VirtualRouterRef

An object that represents a virtual router returned by a list operation.

VirtualRouterServiceProvider

An object that represents a virtual node service provider.

VirtualRouterSpec

An object that represents the specification of a virtual router.

VirtualRouterStatus

An object that represents the status of a virtual router.

VirtualServiceBackend

An object that represents a virtual service backend for a virtual node.

VirtualServiceData

An object that represents a virtual service returned by a describe operation.

VirtualServiceRef

An object that represents a virtual service returned by a list operation.

VirtualServiceSpec

An object that represents the specification of a virtual service.

VirtualServiceStatus

An object that represents the status of a virtual service.

WeightedTarget

An object that represents a target and its relative weight. Traffic is distributed

Enums

AccessLog

Backend

ClientTlsCertificate

DefaultGatewayRouteRewrite

DnsResponseType

DurationUnit

EgressFilterType

GatewayRouteStatusCode

GrpcMetadataMatchMethod

GrpcRetryPolicyEvent

GrpcRouteMetadataMatchMethod

HeaderMatchMethod

HttpMethod

HttpScheme

ListenerTimeout

ListenerTlsCertificate

ListenerTlsMode

ListenerTlsValidationContextTrust

MeshStatusCode

PortProtocol

RouteStatusCode

ServiceDiscovery

TcpRetryPolicyEvent

TlsValidationContextTrust

VirtualGatewayAccessLog

VirtualGatewayClientTlsCertificate

VirtualGatewayConnectionPool

VirtualGatewayListenerTlsCertificate

VirtualGatewayListenerTlsMode

VirtualGatewayListenerTlsValidationContextTrust

VirtualGatewayPortProtocol

VirtualGatewayStatusCode

VirtualGatewayTlsValidationContextTrust

VirtualNodeConnectionPool

VirtualNodeStatusCode

VirtualRouterStatusCode

VirtualServiceProvider

VirtualServiceStatusCode