Module aws_sdk_appmesh::operation

Structs

CreateGatewayRoute

Creates a gateway route.

CreateMesh

Creates a service mesh.

CreateRoute

Creates a route that is associated with a virtual router.

CreateVirtualGateway

Creates a virtual gateway.

CreateVirtualNode

Creates a virtual node within a service mesh.

CreateVirtualRouter

Creates a virtual router within a service mesh.

CreateVirtualService

Creates a virtual service within a service mesh.

DeleteGatewayRoute

Deletes an existing gateway route.

DeleteMesh

Deletes an existing service mesh.

DeleteRoute

Deletes an existing route.

DeleteVirtualGateway

Deletes an existing virtual gateway. You cannot delete a virtual gateway if any gateway

DeleteVirtualNode

Deletes an existing virtual node.

DeleteVirtualRouter

Deletes an existing virtual router.

DeleteVirtualService

Deletes an existing virtual service.

DescribeGatewayRoute

Describes an existing gateway route.

DescribeMesh

Describes an existing service mesh.

DescribeRoute

Describes an existing route.

DescribeVirtualGateway

Describes an existing virtual gateway.

DescribeVirtualNode

Describes an existing virtual node.

DescribeVirtualRouter

Describes an existing virtual router.

DescribeVirtualService

Describes an existing virtual service.

ListGatewayRoutes

Returns a list of existing gateway routes that are associated to a virtual

ListMeshes

Returns a list of existing service meshes.

ListRoutes

Returns a list of existing routes in a service mesh.

ListTagsForResource

List the tags for an App Mesh resource.

ListVirtualGateways

Returns a list of existing virtual gateways in a service mesh.

ListVirtualNodes

Returns a list of existing virtual nodes.

ListVirtualRouters

Returns a list of existing virtual routers in a service mesh.

ListVirtualServices

Returns a list of existing virtual services in a service mesh.

TagResource

Associates the specified tags to a resource with the specified resourceArn.

UntagResource

Deletes specified tags from a resource.

UpdateGatewayRoute

Updates an existing gateway route that is associated to a specified virtual gateway in a

UpdateMesh

Updates an existing service mesh.

UpdateRoute

Updates an existing route for a specified service mesh and virtual router.

UpdateVirtualGateway

Updates an existing virtual gateway in a specified service mesh.

UpdateVirtualNode

Updates an existing virtual node in a specified service mesh.

UpdateVirtualRouter

Updates an existing virtual router in a specified service mesh.

UpdateVirtualService

Updates an existing virtual service in a specified service mesh.